* experimental support for mapping `cMissingValue` descriptors to and from `MayBeMissing<T>` and `Optional<T>` types (both `nil` and `MissingValue` constants will be packed as `cMissingValue` descriptors, while a `cMissingValue` descriptor will unpack as `nil` if the command's return type is explicitly an Optional type, e.g.`String?`, or `MissingValue` if its return type is `Any`)
* removed `Symbol.missingValue` (`cMissingValue` descriptors should never map to Symbols, as that makes distinguishing `missing value` from any other type/enum a PITA and can't take advantage of Swift's type system to bring some degree of predictability/rigour to commands' results)

* moved `formatValue()` function into `FormatterSpecifier` class; this tidies code and allows for future sub-classing of `FormatterSpecifier` in order to change how values are formatted (e.g. to generate syntax for languages other than Swift)

* `aeglue...` comment in glue file now shows arguments in single quotes, no longer includes `-o`/`-r` options

Notes:

Always mapping `missing value` to nil would be the most logical choice, but turns out to be highly problematic as Swift totally sucks at unboxing nils that have been cast to Any. Once an `Optional<T>.none` value (which Swift unwisely treats as a valid member of the Any set) is assigned to a variable of type `Any`, it's damn near impossible to unbox it again unless you _already know_ exactly what T is [so you can cast it back to `Optional<T>` first], which defeats the point of using `Any` in the first place. e.g. The obvious way to unbox it, by casting from `Any` to `Any?`, only results in another layer of boxing being _added, giving the value `Optional.some(nil)`(!) instead of the desired `nil` (a salutary reminder that the road to C*-language usability hell is paved with special-case behaviours all the way down).